+++ tests/testthat/test_subcorpus_bundle.R
@@ -172,3 +172,28 @@ test_that(
     expect_identical(y1, y2)
   }
 )
+
+test_that(
+  "subsetting subcorpus_bundle",
+  {
+    merkel <- corpus("GERMAPARLMINI") %>%
+      split(s_attribute = "protocol_date") %>%
+      subset(speaker == "Angela Dorothea Merkel")
+    expect_identical(length(merkel), 2L)
+
+    skip_on_cran()
+    
+    sp <- corpus("GERMAPARLMINI") %>%
+      as.speeches(
+        s_attribute_name = "speaker",
+        s_attribute_date = "protocol_date"
+      )
+    
+    sp_min <- subset(sp, interjection == "speech", iterate = TRUE, progress = FALSE)
+    expect_identical(length(sp), length(sp_min))
+    expect_identical(
+      unname(unlist(s_attributes(sp_min, "interjection"))),
+      rep("speech", times = length(sp_min))
+    )
+  }
+)